Refactor workspace routes
